feat: migrate config from previous image
Due to the recent rewrite of ublue-os/startingpoint, it made sense
to archive my old uBlue repo and start a new one with the updated
setup.

This migrates some of the old config across, but I havent looked
into what has changed in Silverblue since my last update, so I
anticipate that I'll likely need to troubleshoot a fair bit of
this config before it's suitable for real use again.